backend: Use BigInteger for bytes_freed and size_bytes columns
- Fix MySQL INT overflow when cleanup frees more than ~2.1GB
- Affects cleanup_runs.bytes_freed and cleanup_logs.size_bytes

+++ backend/app/models.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from sqlalchemy import DateTime, Float, Integer, String, Text
+from sqlalchemy import BigInteger, DateTime, Float, Integer, String, Text
 from sqlalchemy.orm import Mapped, mapped_column
 
 from .database import Base
@@ -17,7 +17,7 @@ class CleanupRun(Base):
     disk_usage_before: Mapped[float | None] = mapped_column(Float, nullable=True)
     disk_usage_after: Mapped[float | None] = mapped_column(Float, nullable=True)
     builds_deleted: Mapped[int] = mapped_column(Integer, default=0)
-    bytes_freed: Mapped[int] = mapped_column(Integer, default=0)
+    bytes_freed: Mapped[int] = mapped_column(BigInteger, default=0)
     status: Mapped[str] = mapped_column(String(20), default="running")  # running|completed|failed
     error_message: Mapped[str | None] = mapped_column(Text, nullable=True)
 
@@ -32,6 +32,6 @@ class CleanupLog(Base):
     build_number: Mapped[str] = mapped_column(String(50))
     retention_type: Mapped[str] = mapped_column(String(50))
     age_days: Mapped[float] = mapped_column(Float)
-    size_bytes: Mapped[int] = mapped_column(Integer, default=0)
+    size_bytes: Mapped[int] = mapped_column(BigInteger, default=0)
     score: Mapped[float] = mapped_column(Float)
     dry_run: Mapped[bool] = mapped_column(default=False)
